+++ packages/bpk-docs/src/themeableAttributes.js
@@ -1,0 +1,24 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2017 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/* eslint-disable global-require */
+export default [
+  require('bpk-component-blockquote').themeAttributes,
+  require('bpk-component-button').themeAttributes,
+  require('bpk-component-horizontal-nav').themeAttributes,
+  require('bpk-component-spinner').themeAttributes,
+].filter(attrs => !!attrs);